fix: distinguish between exact and regex matches (#7)
This change proposes to split out the regex matching functionality into
a dedicated `matchRegex` rule. The rationale behind this change is that,
by having `match` support both exact and regex matching it can lead to
very surprising behaviour that is not obvious to the user.

An example: with the presence of regex matching, `match: foo` would
match `foo`, **but it also would match** `foobar` and `barfoo`. This
means that it would match more trust anchors than anticipated, and that
in turn can be a security issue.

To achieve true exact matching, one would need to use `match: ^foo$`,
but that's really verbose and easy to forget.

By making `match` just perform plain exact string matching and have
`matchRegex` for use cases that require more complicated matching
behaviour via regular expressions the footgun above can be avoided.