fix(release): check for `release_created == 'true'` (#32)
The docs of the `release-please-action` state that `release_created` is
`false` if no release was created.

https://github.com/googleapis/release-please-action?tab=readme-ov-file#root-component-outputs

This is incorrect, `release_created` is empty in this case. An empty
string in the `enable` attribute leads to an `Invalid value` error:

https://github.com/Bonial-International-GmbH/sops-check/actions/runs/12117567635/job/33780291842

We'll explicitly compare `release_created` to `'true'` to make sure the
result can only be `true` or `false`, but nothing else.